+++ pkg/controller/namespace/namespace_controller.go
@@ -228,12 +228,7 @@ unzip -o -d "${target_dir}" "${archive}"
 rm -f "${archive}"
 
 echo "Configuring OneAgent..."
-mkdir -p "${target_dir}/agent/conf/pod"
-mkdir -p "${target_dir}/agent/conf/node"
-
 echo -n "${INSTALLPATH}/agent/lib64/liboneagentproc.so" >> "${target_dir}/ld.so.preload"
-echo -n "${NODENAME}" > "${target_dir}/agent/conf/node/name"
-echo -n "${NODEIP}" > "${target_dir}/agent/conf/node/ip"
 `))
 
 func (s *script) generate() (map[string][]byte, error) {